Add SYNCHRONOUS_COMMIT configuration directive disabled by default. This is the current behavior of Ora2Pg, it set synchronous_commit to off before data import to PostgreSQL.
Specifies whether transaction commit will wait for WAL records to be written
to disk before the command returns a "success" indication to the client. This
is the equivalent to set synchronous_commit directive of postgresql.conf file.
This is only used when you load data directly to PostgreSQL, the default is
off to disable synchronous commit to gain speed at writing data. Some modified
version of PostgreSQL, like greenplum, do not have this setting, so in this
set this directive to 1, ora2pg will not  try to change the setting.