llow export of all schemas from an Oracle Instance when SCHEMA directive is empty and EXPORT_SCHEMA is enabled. All exported objects will be prefixed with the name of their original Oracle schema or search_path will be set to that schema name. Thanks to Magnus Hagander for the feature request.
Use COPY FREEZE to export data when COPY_FREEZE is enabled. This will only
works with export to file and when -J or ORACLE_COPIES is not set or default
to 1. It can be used with direct import into PostgreSQL under the same
condition but -j or JOBS must also be unset or default to 1. Thanks to
Magnus Hagander for the feature request.