Add DATA_EXPORT_ORDER configuration directive.
By default data export order will be done by sorting on table name. If you
have huge tables at end of alphabetic order and you are using multiprocess,
it can be better to set the sort order on size so that multiple small tables
can be processed before the largest tables finish. In this case set this
directive to size. Possible values are name and size. Note that export type
SHOW_TABLE and SHOW_COLUMN will use this sort order too, not only COPY or
INSERT export type. Thanks to Guy Browne for the feature request.