- Improve speed of BLOB/CLOB data export by more than 10 times.
Oracle recommends reading from and writing to a LOB in batches using a
multiple of the LOB chunk size. This chunk size defaults to 8k (8192).
Recent tests shown that the best performances can be reach with higher
value like 512K or 4Mb.

A quick benchmark with 30120 rows with different size of BLOB (200x5Mb,
19800x212k, 10000x942K, 100x17Mb, 20x156Mb), with DATA_LIMIT=100,
LONGREADLEN=170Mb and a total table size of 20GB gives:

       no lob locator  : 22m46,218s (1365 sec., avg: 22 recs/sec)
       chunk size 8k   : 15m50,886s (951 sec., avg: 31 recs/sec)
       chunk size 512k : 1m28,161s (88 sec., avg: 342 recs/sec)
       chunk size 4Mb  : 1m23,717s (83 sec., avg: 362 recs/sec)

In conclusion it can be more than 10 time faster with LOB_CHUNK_SIZE set
to 4Mb. Depending of the size of most BLOB you may want to adjust the value
here. For example if you have a majority of small lobs bellow 8K, using 8192
is better to not waste space. By default LOB_CHUNK_SIZE is set to 512000.

- Add progress bar when --oracle_speed is use in single process mode.

- Default for NO_LOB_LOCATOR is now 1 to benefit from the LOB_CHUNK_SIZE
  performances gain. Be warn that this break backward compatibility with
  behavior of previous Ora2Pg versions.