Add new configuration directive ALTERNATIVE_QUOTING_REGEXP to support the Alternative Quoting Mechanism ('Q' or 'q') for String Literals. You can set the regexp with the text capture to use to extract the text part. For example with a variable declared as
        c_sample VARCHAR2(100 CHAR) := q'{This doesn't work.}';

the regexp to use must be:

        ALTERNATIVE_QUOTING_REGEXP     q'{(.*)}'

ora2pg will use the $$ delimiter, with the example the result will be:

        c_sample VARCHAR(100) := $$This doesn't work.$$;

The value of this configuration directive can be a list of regexp
separated by a semi colon. The capture part (between parenthesis) is
mandatory in each regexp if you want to restore the string constant.

Thanks to just-doitfor the report.