Make the build reproducible
Whilst working on the Reproducible Builds effort [0] we noticed that
ora2pg could not be built reproducibly.

This is because it generates a comment in a config file using '$$' in
an attempt to get a literal '$' character. But '$$' inserts the
process PID (!) of the build process into the file instead.

This was originally filed in Debian as #972336 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/972336