Add new option --blob_to_lo that can be used to export BLOB as large objects. It can only be used with action SHOW_COLUMN, TABLE and INSERT. When used with TABLE action, the BLOB column will be translated into oid PostgreSQL data type. When used with the INSERT export action BLOB data will be store as loarge object in the pg_largeobjects table and the oid referencing this large object will be stored in the main table instead of a bytea.
It is not possible to use id with COPY because this feature use function
lo_from_bytea() that stores the large object in the external table and
returns the oid.

This feature works with or without the use of oracle_fdw to import the
data and ORACLE_COPIES or option -J can be used to improve the speed
of the INSERT import provide that there is a numeric unique key on the
table.

Thanks to rodiq for the feature request.