Add new configuration directive PARTITION_BY_REFERENCE to defined How to export Oracle partition by reference. Possible values are none, duplicate or the number of hash partitions to create.
Value 'none' mean no translation and export of partition by reference
like before. Value 'duplicate' will duplicate the referenced column
in the partitioned table and apply the same partitioning from the
referenced table to the partitioned table. If the value is a number,
the table will be partitioned with the HASH method using the value
as the modulo. For example if you set it to 4 it will create 4 HASH
partitions.

Default is none to not export the partitions by reference definition.