Fix incorrect "unsupported partition type" warning for oracle_fdw copy
Using COPY with oracle_fdw on a partitioned table resulted in:

  WARNING: unsupported partition type on table '<table name>'

This is caused by the partition type not being set at the point
export_table() is called to populate $fdw_definition. As "PARTITION BY"
is not applicable to foreign tables the type of partitioning is not
required at this point. This led to the conclusion that the most
appropriate fix is to disable partitioning for the population of
$fdw_definition and then re-establish the original value.